split kinds
