+++ src/packages/core/entity-action/entity-action.extension.ts
@@ -61,137 +61,11 @@ export interface MetaEntityActionDeleteKind extends MetaEntityActionDefaultKind 
 	itemRepositoryAlias: string;
 }
 
-// TRASH
-export interface ManifestEntityActionTrashKind extends ManifestEntityAction<MetaEntityActionTrashKind> {
-	type: 'entityAction';
-	kind: 'trash';
-}
-
-export interface MetaEntityActionTrashKind extends MetaEntityActionDefaultKind {
-	recycleBinRepositoryAlias: string;
-	itemRepositoryAlias: string;
-}
-
-// RESTORE FROM RECYCLE BIN
-
-export interface ManifestEntityActionRestoreFromRecycleBinKind
-	extends ManifestEntityAction<MetaEntityActionRestoreFromRecycleBinKind> {
-	type: 'entityAction';
-	kind: 'restoreFromRecycleBin';
-}
-
-export interface MetaEntityActionRestoreFromRecycleBinKind extends MetaEntityActionDefaultKind {
-	recycleBinRepositoryAlias: string;
-	itemRepositoryAlias: string;
-	pickerModal: UmbModalToken<UmbPickerModalData<any>, UmbPickerModalValue> | string;
-}
-
-// EMPTY RECYCLE BIN
-export interface ManifestEntityActionEmptyRecycleBinKind
-	extends ManifestEntityAction<MetaEntityActionEmptyRecycleBinKind> {
-	type: 'entityAction';
-	kind: 'emptyRecycleBin';
-}
-
-export interface MetaEntityActionEmptyRecycleBinKind extends MetaEntityActionDefaultKind {
-	recycleBinRepositoryAlias: string;
-}
-
-// RENAME
-export interface ManifestEntityActionRenameServerFileKind
-	extends ManifestEntityAction<MetaEntityActionRenameServerFileKind> {
-	type: 'entityAction';
-	kind: 'renameServerFile';
-}
-
-export interface MetaEntityActionRenameServerFileKind extends MetaEntityActionDefaultKind {
-	renameRepositoryAlias: string;
-	itemRepositoryAlias: string;
-}
-
-// RELOAD TREE ITEM CHILDREN
-export interface ManifestEntityActionReloadTreeItemChildrenKind
-	extends ManifestEntityAction<MetaEntityActionRenameServerFileKind> {
-	type: 'entityAction';
-	kind: 'reloadTreeItemChildren';
-}
-
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface MetaEntityActionReloadTreeItemChildrenKind extends MetaEntityActionDefaultKind {}
-
-// DUPLICATE TO
-export interface ManifestEntityActionDuplicateToKind extends ManifestEntityAction<MetaEntityActionDuplicateToKind> {
-	type: 'entityAction';
-	kind: 'duplicateTo';
-}
-
-export interface MetaEntityActionDuplicateToKind extends MetaEntityActionDefaultKind {
-	duplicateRepositoryAlias: string;
-	treeRepositoryAlias: string;
-	treeAlias: string;
-	foldersOnly?: boolean;
-}
-
-// MOVE TO
-export interface ManifestEntityActionMoveToKind extends ManifestEntityAction<MetaEntityActionMoveToKind> {
-	type: 'entityAction';
-	kind: 'moveTo';
-}
-
-export interface MetaEntityActionMoveToKind extends MetaEntityActionDefaultKind {
-	moveRepositoryAlias: string;
-	treeRepositoryAlias: string;
-	treeAlias: string;
-	foldersOnly?: boolean;
-}
-
-// FOLDER
-export interface ManifestEntityActionCreateFolderKind extends ManifestEntityAction<MetaEntityActionFolderKind> {
-	type: 'entityAction';
-	kind: 'folderCreate';
-}
-
-export interface ManifestEntityActionUpdateFolderKind extends ManifestEntityAction<MetaEntityActionFolderKind> {
-	type: 'entityAction';
-	kind: 'folderUpdate';
-}
-
-export interface ManifestEntityActionDeleteFolderKind extends ManifestEntityAction<MetaEntityActionFolderKind> {
-	type: 'entityAction';
-	kind: 'folderDelete';
-}
-
-export interface MetaEntityActionFolderKind extends MetaEntityActionDefaultKind {
-	folderRepositoryAlias: string;
-}
-
-// SORT CHILDREN OF
-export interface ManifestEntityActionSortChildrenOfKind
-	extends ManifestEntityAction<MetaEntityActionSortChildrenOfKind> {
-	type: 'entityAction';
-	kind: 'sortChildrenOf';
-}
-
-export interface MetaEntityActionSortChildrenOfKind extends MetaEntityActionDefaultKind {
-	sortChildrenOfRepositoryAlias: string;
-	treeRepositoryAlias: string;
-}
-
 export type UmbEntityActionExtensions =
 	| ManifestEntityAction
-	| ManifestEntityActionCreateFolderKind
 	| ManifestEntityActionDefaultKind
-	| ManifestEntityActionDeleteFolderKind
 	| ManifestEntityActionDeleteKind
-	| ManifestEntityActionDuplicateToKind
-	| ManifestEntityActionEmptyRecycleBinKind
-	| ManifestEntityActionMoveToKind
-	| ManifestEntityActionReloadTreeItemChildrenKind
-	| ManifestEntityActionRenameServerFileKind
-	| ManifestEntityActionRestoreFromRecycleBinKind
-	| ManifestEntityActionSortChildrenOfKind
-	| ManifestEntityActionTrashKind
-	| ManifestEntityActionUpdateFolderKind;
+	| ManifestEntityActionTrashKind;
 
 /**
  * @deprecated use `UmbEntityActionExtensions` instead.
