+++ src/packages/core/recycle-bin/entity-action/empty-recycle-bin/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './empty-recycle-bin.action.js';
