+++ src/packages/core/recycle-bin/entity-action/index.ts
@@ -1,3 +1,8 @@
-export { UmbTrashEntityAction } from './trash/trash.action.js';
+export { UmbTrashEntityAction } from './trash/index.js';
+export * from './trash/types.js';
+
 export { UmbRestoreFromRecycleBinEntityAction } from './restore-from-recycle-bin/restore-from-recycle-bin.action.js';
+export * from './restore-from-recycle-bin/types.js';
+
 export { UmbEmptyRecycleBinEntityAction } from './empty-recycle-bin/empty-recycle-bin.action.js';
+export * from './empty-recycle-bin/types.js';
