+++ src/packages/core/recycle-bin/entity-action/restore-from-recycle-bin/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './restore-from-recycle-bin.action.js';
