+++ src/packages/core/recycle-bin/entity-action/trash/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './trash.action.js';
