+++ src/packages/core/recycle-bin/entity-action/trash/types.ts
@@ -1,0 +1,17 @@
+import type { ManifestEntityAction, MetaEntityActionDefaultKind } from '@umbraco-cms/backoffice/entity-action';
+
+export interface ManifestEntityActionTrashKind extends ManifestEntityAction<MetaEntityActionTrashKind> {
+	type: 'entityAction';
+	kind: 'trash';
+}
+
+export interface MetaEntityActionTrashKind extends MetaEntityActionDefaultKind {
+	recycleBinRepositoryAlias: string;
+	itemRepositoryAlias: string;
+}
+
+declare global {
+	interface UmbExtensionManifestMap {
+		umbTrashEntityActionKind: ManifestEntityActionTrashKind;
+	}
+}
