+++ src/packages/core/recycle-bin/index.ts
@@ -1,10 +1,4 @@
 export { UmbRecycleBinRepositoryBase } from './recycle-bin-repository-base.js';
-export {
-	UmbTrashEntityAction,
-	UmbRestoreFromRecycleBinEntityAction,
-	UmbEmptyRecycleBinEntityAction,
-} from './entity-action/index.js';
-
 export type { UmbRecycleBinDataSource } from './recycle-bin-data-source.interface.js';
 export type { UmbRecycleBinRepository } from './recycle-bin-repository.interface.js';
 export type {
@@ -12,6 +6,8 @@ export type {
 	UmbRecycleBinTrashRequestArgs,
 	UmbRecycleBinOriginalParentRequestArgs,
 } from './types.js';
+
+export * from './entity-action/index.js';
 
 export { UmbIsTrashedEntityContext, UMB_IS_TRASHED_ENTITY_CONTEXT } from './contexts/is-trashed/index.js';
 export { UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS } from './conditions/is-not-trashed/constants.js';
