+++ src/packages/core/server-file-system/rename/types.ts
@@ -1,5 +1,6 @@
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbEntityModel } from '@umbraco-cms/backoffice/entity';
+import type { ManifestEntityAction, MetaEntityActionDefaultKind } from '@umbraco-cms/backoffice/entity-action';
 import type { UmbDataSourceResponse } from '@umbraco-cms/backoffice/repository';
 
 export interface UmbRenameServerFileRepository<DetailType extends UmbEntityModel> {
@@ -12,4 +13,21 @@ export interface UmbRenameServerFileDataSourceConstructor<DetailType extends Umb
 
 export interface UmbRenameServerFileDataSource<DetailType extends UmbEntityModel> {
 	rename(unique: string, name: string): Promise<UmbDataSourceResponse<DetailType>>;
+}
+
+export interface ManifestEntityActionRenameServerFileKind
+	extends ManifestEntityAction<MetaEntityActionRenameServerFileKind> {
+	type: 'entityAction';
+	kind: 'renameServerFile';
+}
+
+export interface MetaEntityActionRenameServerFileKind extends MetaEntityActionDefaultKind {
+	renameRepositoryAlias: string;
+	itemRepositoryAlias: string;
+}
+
+declare global {
+	interface UmbExtensionManifestMap {
+		umbRenameFileServerFileEntityActionKind: ManifestEntityActionRenameServerFileKind;
+	}
 }
