+++ src/packages/core/tree/entity-actions/duplicate-to/duplicate-to.action.ts
@@ -1,10 +1,6 @@
 import { UMB_DUPLICATE_TO_MODAL } from './modal/duplicate-to-modal.token.js';
-import type { UmbDuplicateToRepository } from './types.js';
-import {
-	UmbEntityActionBase,
-	UmbRequestReloadStructureForEntityEvent,
-	type MetaEntityActionDuplicateToKind,
-} from '@umbraco-cms/backoffice/entity-action';
+import type { MetaEntityActionDuplicateToKind, UmbDuplicateToRepository } from './types.js';
+import { UmbEntityActionBase, UmbRequestReloadStructureForEntityEvent } from '@umbraco-cms/backoffice/entity-action';
 import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 import { createExtensionApiByAlias } from '@umbraco-cms/backoffice/extension-registry';
 import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
