+++ src/packages/core/tree/entity-actions/duplicate-to/types.ts
@@ -1,4 +1,8 @@
-import type { UmbDuplicateRequestArgs } from '@umbraco-cms/backoffice/entity-action';
+import type {
+	ManifestEntityAction,
+	MetaEntityActionDefaultKind,
+	UmbDuplicateRequestArgs,
+} from '@umbraco-cms/backoffice/entity-action';
 
 export * from './duplicate-to-data-source.interface.js';
 export * from './duplicate-to-repository.interface.js';
@@ -7,4 +11,22 @@ export interface UmbDuplicateToRequestArgs extends UmbDuplicateRequestArgs {
 	destination: {
 		unique: string | null;
 	};
+}
+
+export interface ManifestEntityActionDuplicateToKind extends ManifestEntityAction<MetaEntityActionDuplicateToKind> {
+	type: 'entityAction';
+	kind: 'duplicateTo';
+}
+
+export interface MetaEntityActionDuplicateToKind extends MetaEntityActionDefaultKind {
+	duplicateRepositoryAlias: string;
+	treeRepositoryAlias: string;
+	treeAlias: string;
+	foldersOnly?: boolean;
+}
+
+declare global {
+	interface UmbExtensionManifestMap {
+		umbDuplicateToEntityActionKind: ManifestEntityActionDuplicateToKind;
+	}
 }
