+++ src/packages/core/tree/entity-actions/move/index.ts
@@ -1,3 +1,3 @@
 export type { UmbMoveDataSource } from './move-data-source.interface.js';
 export type { UmbMoveRepository } from './move-repository.interface.js';
-export type { UmbMoveToRequestArgs } from './types.js';
+export * from './types.js';
