+++ src/packages/core/tree/entity-actions/reload-tree-item-children/types.ts
@@ -1,0 +1,16 @@
+import type { ManifestEntityAction, MetaEntityActionDefaultKind } from '@umbraco-cms/backoffice/entity-action';
+
+export interface ManifestEntityActionReloadTreeItemChildrenKind
+	extends ManifestEntityAction<MetaEntityActionReloadTreeItemChildrenKind> {
+	type: 'entityAction';
+	kind: 'reloadTreeItemChildren';
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type
+export interface MetaEntityActionReloadTreeItemChildrenKind extends MetaEntityActionDefaultKind {}
+
+declare global {
+	interface UmbExtensionManifestMap {
+		umbReloadTreeItemChildrenEntityAction: ManifestEntityActionReloadTreeItemChildrenKind;
+	}
+}
