+++ src/packages/core/tree/entity-actions/sort-children-of/index.ts
@@ -1,4 +1,4 @@
 export { UmbSortChildrenOfEntityAction } from './sort-children-of.action.js';
-export type { UmbSortChildrenOfArgs } from './types.js';
 export type { UmbSortChildrenOfRepository } from './sort-children-of-repository.interface.js';
 export type { UmbSortChildrenOfDataSource } from './sort-children-of-data-source.interface.js';
+export * from './types.js';
