+++ src/packages/core/tree/folder/entity-action/create-folder/create-folder.action.ts
@@ -1,7 +1,7 @@
 import { UMB_FOLDER_CREATE_MODAL } from '../../modal/index.js';
+import type { MetaEntityActionFolderKind } from '../../types.js';
 import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
 import { UmbEntityActionBase, UmbRequestReloadChildrenOfEntityEvent } from '@umbraco-cms/backoffice/entity-action';
-import type { MetaEntityActionFolderKind } from '@umbraco-cms/backoffice/entity-action';
 import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 
 export class UmbCreateFolderEntityAction extends UmbEntityActionBase<MetaEntityActionFolderKind> {
