+++ src/packages/core/tree/folder/entity-action/delete-folder/index.ts
@@ -1,0 +1,2 @@
+export * from './delete-folder.action.js';
+export * from './types.js';
