+++ src/packages/core/tree/folder/entity-action/update-folder/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './update-folder.action.js';
