+++ src/packages/core/tree/folder/entity-action/update-folder/update-folder.action.ts
@@ -1,7 +1,7 @@
 import { UMB_FOLDER_UPDATE_MODAL } from '../../modal/index.js';
+import type { MetaEntityActionFolderKind } from '../../types.js';
 import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
 import { UmbEntityActionBase, UmbRequestReloadStructureForEntityEvent } from '@umbraco-cms/backoffice/entity-action';
-import type { MetaEntityActionFolderKind } from '@umbraco-cms/backoffice/entity-action';
 import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 
 export class UmbUpdateFolderEntityAction extends UmbEntityActionBase<MetaEntityActionFolderKind> {
